requirements.txt
Django==1.7.1
Pillow==3.0.0
python-gcm==0.3
pytz==2015.7
qrcode==5.1
requests==2.8.1
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
